+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Speed Scale for distance sequencers, turning them into arpeggiators that keep pace with the listener
+- Sequencer tracks can be modulated from a sound element, like sound elements already could
 
 ### Changed